Send book request via Socket instead of deprecated AsyncTask
This change is not motivated by a failure I or anyone else has observed. But using deprecated code is not good long term, so this change replaces AsyncTask (deprecated in API level 30) with a Socket. This also has the minor benefit of reducing the thread count by 1.
Some debug output is still present. I will remove it before making a PR.